♻️: shared/ui Table 컴포넌트 리팩토링 - 분리
- Table 컴포넌트
- TableBody 컴포넌트
- TableCell 컴포넌트
- TableHead 컴포넌트
- TableHeader 컴포넌트
- TableRow 컴포넌트
